# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: VJU Club Management API
  version: 1.0.0
  description: 'API giai đoạn 1 cho quản lý tài khoản, CLB, ban, thành viên và permission.
    Xem x-implementation-status từng operation: IMPLEMENTED là đã xác nhận qua integration
    test. UUID, camelCase, thời gian ISO-8601. List dùng offset/limit; không có page/size.'
servers:
- url: /api/v1
  description: Backend cùng origin với Swagger UI
- url: http://localhost:8080/api/v1
  description: Backend local mặc định
security:
- sessionAuth: []
paths:
...
  /users/me:
    get:
      operationId: user.me
      summary: Get current profile
      tags:
      - User
      x-permission: null
      x-implementation-status: IMPLEMENTED
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              example:
                id: 11111111-1111-4111-8111-111111111111
                email: phat@example.test
                fullName: Nguyễn Xuân Phát
                studentCode: '23110209'
                phone: 0900000000
                avatarUrl: null
                status: ACTIVE
                createdAt: '2026-09-29T12:00:00Z'
                updatedAt: '2026-09-29T12:00:00Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequest'
      description: Chỉ đọc profile của tài khoản hiện tại. Không cần user.view.
    patch:
      operationId: user.updateMe
      summary: Update current profile
      tags:
      - User
      x-permission: null
      x-implementation-status: IMPLEMENTED
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              example:
                id: 11111111-1111-4111-8111-111111111111
                email: phat@example.test
                fullName: Nguyễn Xuân Phát
                studentCode: '23110209'
                phone: 0900000000
                avatarUrl: null
                status: ACTIVE
                createdAt: '2026-09-29T12:00:00Z'
                updatedAt: '2026-09-29T12:00:00Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
            example:
              fullName: Nguyễn Xuân Phát
              phone: 0900000000
      description: Cập nhật fullName, phone, avatarUrl của tài khoản hiện tại. Không
        cho sửa status, permission hay membership. Field bỏ qua/null giữ nguyên.
  /users:
    post:
      operationId: user.create
      summary: Create account (admin)
      tags:
      - User
      x-permission: user.create
      x-implementation-status: IMPLEMENTED
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              example:
                id: 11111111-1111-4111-8111-111111111111
                email: phat@example.test
                fullName: Nguyễn Xuân Phát
                studentCode: '23110209'
                phone: 0900000000
                avatarUrl: null
                status: ACTIVE
                createdAt: '2026-09-29T12:00:00Z'
                updatedAt: '2026-09-29T12:00:00Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '400':
          $ref: '#/components/responses/BadRequest'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
            example:
              email: phat@example.test
              password: DemoPassword123!
              fullName: Nguyễn Xuân Phát
              studentCode: '23110209'
              phone: 0900000000
      description: Chỉ admin (quyền user.create) tạo tài khoản; hệ thống không cho tự đăng ký.
        Tài khoản tạo ra ở trạng thái ACTIVE, chưa có permission hay membership. Mật khẩu được hash. Email trùng trả 409.
    get:
      operationId: user.list
      summary: List users
      tags:
      - User
      x-permission: user.view
      x-implementation-status: IMPLEMENTED
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
              example:
                items:
                - id: 11111111-1111-4111-8111-111111111111
                  email: phat@example.test
                  fullName: Nguyễn Xuân Phát
                  studentCode: '23110209'
                  phone: 0900000000
                  avatarUrl: null
                  status: ACTIVE
                  createdAt: '2026-09-29T12:00:00Z'
                  updatedAt: '2026-09-29T12:00:00Z'
                total: 1
                offset: 0
                limit: 20
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequest'
      parameters:
      - name: query
        in: query
        schema:
          type: string
          default: ''
        description: Tìm theo tên/mã CLB hoặc email/họ tên/mã sinh viên tùy endpoint.
      - name: offset
        in: query
        schema:
          type: integer
          minimum: 0
          default: 0
        description: Số bản ghi bỏ qua; bắt đầu từ 0.
      - name: limit
        in: query
        schema:
          type: integer
          minimum: 1
          default: 20
          maximum: 100
        description: Số bản ghi tối đa trong một lần lấy, mặc định 20, tối đa 100.
      - name: orderBy
        in: query
        description: Cột sắp xếp được backend cho phép.
        schema:
          type: string
          enum:
          - email
          - fullName
          - createdAt
          - status
          default: createdAt
      - name: orderType
        in: query
        schema:
          type: string
          enum:
          - asc
          - desc
          default: desc
  /users/{userId}:
    get:
      operationId: user.get
      summary: Get user
      tags:
      - User
      x-permission: user.view
      x-implementation-status: IMPLEMENTED
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              example:
                id: 11111111-1111-4111-8111-111111111111
                email: phat@example.test
                fullName: Nguyễn Xuân Phát
                studentCode: '23110209'
                phone: 0900000000
                avatarUrl: null
                status: ACTIVE
                createdAt: '2026-09-29T12:00:00Z'
                updatedAt: '2026-09-29T12:00:00Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
  /users/{userId}/password:
    patch:
      operationId: user.reset_password
      summary: Reset a user password (admin)
      tags:
      - User
      x-permission: user.reset_password
      x-implementation-status: IMPLEMENTED
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '204':
          description: Password changed; every refresh token of the user is revoked
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      description: 'Người dùng quên mật khẩu liên hệ admin; admin đặt mật khẩu mới. Không có API tự đăng ký hay tự khôi phục mật khẩu.'
  /users/{userId}/status:
    patch:
      operationId: user.status
      summary: Update user status
      tags:
      - User
      x-permission: user.active | user.inactive
      x-implementation-status: IMPLEMENTED
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              example:
                id: 11111111-1111-4111-8111-111111111111
                email: phat@example.test
                fullName: Nguyễn Xuân Phát
                studentCode: '23110209'
                phone: 0900000000
                avatarUrl: null
                status: ACTIVE
                createdAt: '2026-09-29T12:00:00Z'
                updatedAt: '2026-09-29T12:00:00Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserStatusRequest'
            example:
              status: INACTIVE
      x-permission-rules:
        ACTIVE: user.active
        INACTIVE: user.inactive
      description: 'Quyền phụ thuộc trạng thái mới trong body: ACTIVE → user.active,
        INACTIVE → user.inactive.'
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: CLUB_SESSION
      description: 'Cookie phiên do POST /auth/login đặt. Request thay đổi dữ liệu cần thêm header X-XSRF-TOKEN.'
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: 'Chưa đăng nhập hoặc phiên đã hết hạn/bị huỷ (code UNAUTHORIZED); đăng nhập lại.'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: Permission denied
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: Trùng dữ liệu hoặc trạng thái hiện tại không cho phép thao tác.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Không tìm thấy resource hoặc grant.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    CreateUserRequest:
      type: object
      required:
      - email
      - password
      - fullName
      properties:
        email:
          type: string
          format: email
          maxLength: 320
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 72
          description: '8+ ký tự, tối đa 72 byte UTF-8 (giới hạn của BCrypt; chữ có dấu tốn 2–3 byte)'
        fullName:
          type: string
          maxLength: 200
        studentCode:
          type:
          - string
          - 'null'
          maxLength: 50
        phone:
          type:
          - string
          - 'null'
          maxLength: 32
          pattern: '^[0-9+().\- ]*$'
      examples:
      - email: phat@example.test
        password: DemoPassword123!
        fullName: Nguyễn Xuân Phát
        studentCode: '23110209'
        phone: 0900000000
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        email:
          type: string
          format: email
        fullName:
          type: string
        studentCode:
          type:
          - string
          - 'null'
        phone:
          type:
          - string
          - 'null'
        avatarUrl:
          type:
          - string
          - 'null'
        status:
          type: string
          enum:
          - ACTIVE
          - INACTIVE
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
      required:
      - id
      - email
      - fullName
      - studentCode
      - phone
      - avatarUrl
      - status
      - createdAt
      - updatedAt
      examples:
      - id: 11111111-1111-4111-8111-111111111111
        email: phat@example.test
        fullName: Nguyễn Xuân Phát
        studentCode: '23110209'
        phone: 0900000000
        avatarUrl: null
        status: ACTIVE
        createdAt: '2026-09-29T12:00:00Z'
        updatedAt: '2026-09-29T12:00:00Z'
    UpdateProfileRequest:
      type: object
      properties:
        fullName:
          type: string
          maxLength: 200
        phone:
          type:
          - string
          - 'null'
          maxLength: 32
          pattern: '^[0-9+().\- ]*$'
        avatarUrl:
          type:
          - string
          - 'null'
          maxLength: 500
          format: uri
          pattern: '^(https?://.+)?$'
      examples:
      - fullName: Nguyễn Xuân Phát
        phone: 0900000000
    UpdateUserStatusRequest:
      type: object
      required:
      - status
      properties:
        status:
          type: string
          enum:
          - ACTIVE
          - INACTIVE
      examples:
      - status: INACTIVE
...
    Problem:
      type: object
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
          description: URI request, có thể vắng với lỗi do security filter trả về.
        code:
          type: string
      required:
      - type
      - title
      - status
      - detail
      - code
      examples:
      - type: about:blank
        title: PERMISSION_DENIED
        status: 403
        detail: Permission denied
        instance: /api/v1/clubs/22222222-2222-4222-8222-222222222222
        code: PERMISSION_DENIED
    UserListResponse:
      type: object
      required:
      - items
      - total
      - offset
      - limit
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/User'
        total:
          type: integer
          format: int64
        offset:
          type: integer
        limit:
          type: integer
      examples:
      - items:
        - id: 11111111-1111-4111-8111-111111111111
          email: phat@example.test
          fullName: Nguyễn Xuân Phát
          studentCode: '23110209'
          phone: 0900000000
          avatarUrl: null
          status: ACTIVE
          createdAt: '2026-09-29T12:00:00Z'
          updatedAt: '2026-09-29T12:00:00Z'
        total: 1
        offset: 0
        limit: 20
...
    ResetPasswordRequest:
      type: object
      required: [newPassword]
      properties:
        newPassword: {type: string, minLength: 8}
...